#104 feat: select de status funcionando
